services/barber/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
apscheduler==3.10.4
google-auth==2.35.0
google-api-python-client==2.149.0
paho-mqtt==2.1.0
python-dotenv==1.0.1
